requirements.txt
pinecone-client
openai
pymongo
flask
flask-cors
Flask-HTTPAuth